backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
mysql-connector-python==9.1.0
requests==2.32.3
openai==1.52.2
anthropic==0.37.1
alembic==1.13.3